Create PD_testing.py
Precision and Recall are used to calculate F1 score to compare models with
different window size and gamma values.Recall is the number of true positives divided by the
number of true positives plus the number of false negatives. Precision refers to the number of
true positives divided by the number of true positives plus the number of false positives.
Receiver Operating Curve is plotted and area under the curve is calculated for all the models
trained. A great pedestrian detection system will have an ROC curve that is very close to upper
left corner of the plot.